refactor(sensor): split test_telemetry_native.cpp into per-domain test files
Split the 1717-line test file into 11 focused files: test_fakes (shared
fake collectors), test_helpers (utilities), and per-domain test files
for system_snapshot, process, disk, network, thermal, per_core_cpu, gpu,
and process_management. Slimmed runner forward-declares all 42 tests.
No behavioral changes — pure file reorganization.
